Reorganize directories.
