+++ lib/src/embedded-compiler/compiler.test.ts
