+++ lib/src/embedded-protocol/dispatcher.test.ts
