+++ lib/src/embedded-protocol/dispatcher.ts
