+++ lib/src/embedded-protocol/message-transformer.test.ts
