+++ lib/src/embedded-protocol/message-transformer.ts
